Temporary solution for TypeScript
